requirements.txt
absl-py==0.8.0
argh==0.26.2
astor==0.8.0
backports.tempfile==1.0
backports.weakref==1.0.post1
cffi==1.12.3
chardet==3.0.4
Click==7.0
cycler==0.10.0
Cython==0.29.13
decorator==4.4.0
docker-pycreds==0.4.0
future==0.17.1
gast==0.3.2
gitdb2==2.0.6
GitPython==3.0.2
glfw==1.8.3
google-pasta==0.1.7
gql==0.1.0
graphql-core==2.2.1
grpcio==1.24.0
gym==0.10.8
h5py==2.10.0
idna==2.8
imageio==2.4.1
joblib==0.13.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
libtmux==0.8.1
lockfile==0.12.2
Markdown==3.1.1
matplotlib==3.1.2
moviepy==1.0.1
mujoco-py==2.0.2.5
numpy==1.15.2
nvidia-ml-py3==7.352.0
opencv-python==4.0.0.21
pandas==0.24.2
pathtools==0.1.2
Pillow==6.1.0
promise==2.2.1
protobuf==3.9.2
psutil==5.6.3
pycparser==2.19
pygame==1.9.6
pyglet==1.4.4
pymunk==5.4.2
pyparsing==2.4.2
PyQt5==5.13.2
PyQt5-sip==12.7.0
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
Rx==1.6.1
scikit-learn==0.21.3
scipy==1.3.1
seaborn==0.10.0
sentry-sdk==0.12.2
shortuuid==0.5.0
six==1.12.0
sklearn==0.0
smmap2==2.0.5
subprocess32==3.5.4
tensorboard==1.14.0
tensorboardX==1.8
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
torch==1.2.0
tqdm==4.36.1
urllib3==1.25.6
wandb==0.8.3
watchdog==0.9.0
Werkzeug==0.16.0
wrapt==1.11.2
